refactor stucture of veiws and models, updating urls, adding get, list, create defined endpoints
