+++ .isort.cfg
@@ -7,7 +7,7 @@ import_heading_thirdparty=Third-Party Libraries
 import_heading_firstparty=cisagov Libraries
 
 # Should be auto-populated by seed-isort-config hook
-known_third_party=OpenSSL,bson,celery,compat,config,cryptography,database,datrie,django,genshi,gophish,lxml,motor,ntlm,pip,pkg_resources,reports,requests,rest_framework,schematics,service,setuptools,test_validators,weasyprint,yaml
+known_third_party=OpenSSL,api,bson,celery,compat,config,cryptography,database,datrie,django,genshi,gophish,lxml,motor,ntlm,pip,pkg_resources,reports,requests,rest_framework,schematics,service,setuptools,test_validators,weasyprint,yaml
 # These must be manually set to correctly separate them from third party libraries
 known_first_party=
 # This helps black and isort get along
