+++ controller/src/api/urls.py
@@ -1,9 +1,21 @@
-from django.contrib import admin
-from django.urls import path, include
+"""
+API URLs.
 
-from . import views
-
+This lists all urls under the API app.
+"""
+# Third-Party Libraries
+from api.views import subscription_views
+from django.urls import path
 
 urlpatterns = [
-     path("v1/subscriptions/", views.SubscriptionsView.as_view(), name="subscriptions_api")
+    path(
+        "v1/subscriptions/",
+        subscription_views.SubscriptionsListView.as_view(),
+        name="subscriptions_list_api",
+    ),
+    path(
+        "v1/subscription/<subscription_uuid>/",
+        subscription_views.SubscriptionView.as_view(),
+        name="subscriptions_get_api",
+    ),
 ]
